service/requirements.txt
Django==3.2.11
djangorestframework~=3.13.1
django-cors-headers==3.10.1
psycopg2-binary==2.8.6
django-filter==21.1
